feat(image): merge into one file
